Carousel fixes

+++ Growler/Components/SwiftCarousel/SwiftCarouselResizeType.swift
@@ -30,7 +30,7 @@ public enum SwiftCarouselResizeType {
     /// VisibleItemsPerPage will try to fit the number of items you specify
     /// in the whole screen (will resize them of course).
     /// Parameter = number of items visible on screen.
-    case visibleItemsPerPage(CGFloat)
+    case visibleItemsPerPage(Int)
     
     /// FloatWithSpacing will use sizeToFit() on your views to correctly place images
     /// It is helpful for instance with UILabels (Example1 in Examples folder).
